Remove unused image-bumper binary and move shared library to Prow.
