+++ prow/cmd/generic-autobumper/imagebumper/imagebumper.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package bumper
+package imagebumper
 
 import (
 	"encoding/json"
